requirements.txt
mss==9.0.1
Flask==3.0.0
Pillow==10.1.0
PyYAML==6.0.1
ImageHash==4.3.1
requests>=2.31.0
pynput>=1.7.6
python-dateutil>=2.8.2
pytest==7.4.3
pytest-cov==4.1.0